Move checkout-pages to /checkout
