+++ src/data/texts/texts.no.js
@@ -110,19 +110,19 @@ const textsNO = {
       TITLE: "Kjøpsvilkår",
     },
     CHECKOUT_CART: {
-      URL: "/no/butikk/handlekurv",
+      URL: "/no/butikk/kassen/handlekurv",
       TITLE: "Handlekurv",
     },
     CHECKOUT_PERSONAL_DETAILS: {
-      URL: "/no/butikk/personlige-opplysninger",
+      URL: "/no/butikk/kassen/personlige-opplysninger",
       TITLE: "Personlige Opplysninger",
     },
     CHECKOUT_PAYMENT_DETAILS: {
-      URL: "/no/butikk/betaling",
+      URL: "/no/butikk/kassen/betaling",
       TITLE: "Betaling",
     },
     CHECKOUT_RECEIPT: {
-      URL: "/no/butikk/kvittering",
+      URL: "/no/butikk/kassen/kvittering",
       TITLE: "Kvittering",
     },
   },
