+++ src/pages/butikk/kassen/betaling.no.js
@@ -1,8 +1,8 @@
 import React from "react"
 import * as PropTypes from "prop-types"
-import App from "../../components/app"
-import TEXTS from "../../data/texts/texts.no.js"
-import PaymentDetailsPage from "../../page-templates/payment-details-page"
+import App from "../../../components/app"
+import TEXTS from "../../../data/texts/texts.no.js"
+import PaymentDetailsPage from "../../../page-templates/payment-details-page"
 
 const PaymentDetails = ({ location }) => {
   return (
