+++ src/pages/butikk/kassen/handlekurv.no.js
@@ -1,9 +1,9 @@
 import React from "react"
 import * as PropTypes from "prop-types"
 
-import App from "../../components/app"
-import TEXTS from "../../data/texts/texts.no"
-import CartPage from "../../page-templates/cart-page"
+import App from "../../../components/app"
+import TEXTS from "../../../data/texts/texts.no"
+import CartPage from "../../../page-templates/cart-page"
 
 const Handlekurv = ({ location }) => (
   <App location={location} pageTitle={TEXTS.CART}>
