+++ src/pages/butikk/kassen/personlige-opplysninger.no.js
@@ -1,9 +1,9 @@
 import React from "react"
 import * as PropTypes from "prop-types"
 
-import App from "../../components/app"
-import TEXTS from "../../data/texts/texts.no"
-import PersonalDetailsPage from "../../page-templates/personal-details-page"
+import App from "../../../components/app"
+import TEXTS from "../../../data/texts/texts.no"
+import PersonalDetailsPage from "../../../page-templates/personal-details-page"
 
 const PersonalDetails = ({ location }) => {
   return (
