+++ src/pages/shop/checkout/receipt.en.js
@@ -1,9 +1,9 @@
 import React from "react"
 import * as PropTypes from "prop-types"
 
-import App from "../../components/app"
-import ReceiptPage from "../../page-templates/receipt-page"
-import TEXTS from "../../data/texts/texts.en"
+import App from "../../../components/app"
+import ReceiptPage from "../../../page-templates/receipt-page"
+import TEXTS from "../../../data/texts/texts.en"
 
 const Receipt = ({ location }) => {
   return (
